Start of configuration overhaul
In preparation for a SimilarityIndex server, which requires SI
configuration encapsulation. It is now the case that constructors are
using more complex arguments, necessitating a method of letting the
implementation class optionally do what it wants with its configuration
dictionary rather than adhering to a strict external standard.

This should also allow for nested plugin configurations which we will
see in the ITQ configuration when parsing for its CodeIndex factory.

This commit also includes the start of the SimilarityIndex update, which
removes the save/load methods, shifting persistant storage parameters
into class constructors, which also causes those implementations that
can store persistantly to have to store data. This also allows for
construction of an index with a pre-built index instead of having to
explicitly call a load method. This allows us to simply specify a
construction configuration that directly maps to the constructor and
specified an existing index.