Update plugin discovery logging when checking module attributes
Before, a multi-part if-statment determined if a module attribute was
skipped for inclusion as an interface implementation but no reasons were
output as to why.  Broke apart the large if-check into parts so specific
logging occurs to explain why attributes are not kept.  Prevented the
logging from being over verbose by gating descriptive output to only
module attributes that are python types (i.e. a function cannot
implement an abstract interface, etc.).