Fix incorrect logger usage in plugins util
Was incorrectly calling the module level debug funciton instead of the
logger instance that was passed to the function. This caused an
``logging.basicConfig()`` to be called, which would always cause a
Streamhandler to be added to the root logger, causing duplicate logging
statements when specific streamhandlers are added to sub-namespaces.
This fix switches to using the logger instance passed to the function.