Learned that derrived classes of Configurable need new type annotation
Learned new pattern required for derriving classes that descend from
Configurable in that they ought to use a new TypeVar definition instead
of using the one already defined in `smqtk_core.configuration`.  The
`from_config` classmethod documentation and in-line example have been
updated to reflect this with an explanation on why.