Adjust git conflict marker length for .rst files
It turns out that there are valid cases where 7-character long sequnces
of the `=` symbol can validly come up. This is the default length for
git's conflict markers. A check that ghostflow performs is roughly `git
diff --check master`, which will error if it finds changes that
introduce conflict markers or whitespace errors. Since `=======` may be
valid in RST files for some lengths of sections, we adjust the conflict
marker length to the opposite end of the spectrum of "very long" (79).