Experimental revision to setup
This is an attempt to reduce complexity and make better use of
`requirements.txt` file to track a full development environment instead
of just top level topical requirements.

*Discussion:*
Previously we had separated out requirements into different
`requirements.txt` files, then reading those files in during the
execution of `setup.py`, stripping versions appropriately in order to
satisfy the purpose of installation requirements to be abstract as
opposed to concrete. However, I have been uncomfortable with that
version stripping logic being effectively a duplication of similar
parsing logic hidden within the private API of pip. This is compounded
with the existance of this duplicated logic being an efford to "reduce"
duplication of depedency specification between the "install_required"
list and the `requirements.txt` list. This could be a problem that is
already "solved" by project helper tools like `pipenv` and `poetry`.

One subtle gap that I still see in the use of pipenv or poetry
is that it is relatively easy to develop under the context of what the
tool installs for you, which tends to include developer tools, causing
an increased risk of accidentally creating dependencies within the
package proper on the development tools installed, which we want to
avoid.